# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST for LiveCart SaaS — live commerce order detection, cart consolidation, and integrations.",
        "title": "LiveCart API",
        "termsOfService": "https://livecart.com/terms",
        "contact": {
            "name": "LiveCart Support",
            "email": "support@livecart.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "paths": {
...
        "/api/v1/stores": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new store with the given name and slug",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stores"
                ],
                "summary": "Create a new store",
                "parameters": [
                    {
                        "description": "Store creation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apps_api_internal_store.CreateStoreRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apps_api_internal_store.CreateStoreResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.ValidationEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/stores/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the store associated with the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stores"
                ],
                "summary": "Get current store",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apps_api_internal_store.StoreResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the store associated with the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stores"
                ],
                "summary": "Update current store",
                "parameters": [
                    {
                        "description": "Store update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apps_api_internal_store.UpdateStoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apps_api_internal_store.StoreResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.ValidationEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/stores/me/cart-settings": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the cart settings for the authenticated user's store",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stores"
                ],
                "summary": "Update cart settings",
                "parameters": [
                    {
                        "description": "Cart settings payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apps_api_internal_store.UpdateCartSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apps_api_internal_store.StoreResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.ValidationEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/stores/{storeId}/cart-settings": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the cart settings for a specific store (requires store access)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stores"
                ],
                "summary": "Update cart settings for a store",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store UUID",
                        "name": "storeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cart settings payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apps_api_internal_store.UpdateCartSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apps_api_internal_store.StoreResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.ValidationEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/stores/{storeId}/complete-onboarding": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks the store's onboarding as complete",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stores"
                ],
                "summary": "Complete store onboarding",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store UUID",
                        "name": "storeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/me/stores": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all stores the authenticated user belongs to",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all stores for current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apps_api_internal_user.GetUserStoresResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/livecart_apps_api_lib_httpx.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apps_api_internal_store.CartSettingsDTO": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "expirationMinutes": {
                    "type": "integer"
                },
                "maxItems": {
                    "type": "integer"
                },
                "maxQuantityPerItem": {
                    "type": "integer"
                },
                "notifyBeforeExpiration": {
                    "type": "boolean"
                },
                "reserveStock": {
                    "type": "boolean"
                }
            }
        },
        "apps_api_internal_store.CreateStoreRequest": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "slug": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                }
            }
        },
        "apps_api_internal_store.CreateStoreResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "apps_api_internal_store.StoreResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "cartSettings": {
                    "$ref": "#/definitions/apps_api_internal_store.CartSettingsDTO"
                },
                "createdAt": {
                    "type": "string"
                },
                "emailAddress": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "smsNumber": {
                    "type": "string"
                },
                "whatsappNumber": {
                    "type": "string"
                }
            }
        },
        "apps_api_internal_store.UpdateCartSettingsRequest": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "expirationMinutes": {
                    "type": "integer",
                    "minimum": 0
                },
                "maxItems": {
                    "type": "integer",
                    "minimum": 0
                },
                "maxQuantityPerItem": {
                    "type": "integer",
                    "minimum": 0
                },
                "notifyBeforeExpiration": {
                    "type": "boolean"
                },
                "reserveStock": {
                    "type": "boolean"
                }
            }
        },
        "apps_api_internal_store.UpdateStoreRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "emailAddress": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "smsNumber": {
                    "type": "string"
                },
                "whatsappNumber": {
                    "type": "string"
                }
            }
        },
...
        "apps_api_internal_user.GetUserStoresResponse": {
            "type": "object",
            "properties": {
                "stores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apps_api_internal_user.UserStoreResponse"
                    }
                }
            }
        },
...
        "apps_api_internal_user.UserStoreResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "storeId": {
                    "type": "string"
                },
                "storeName": {
                    "type": "string"
                },
                "storeSlug": {
                    "type": "string"
                }
            }
        },
...
        "livecart_apps_api_lib_httpx.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "livecart_apps_api_lib_httpx.ValidationEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}